Use rects in config to figure out SN and EPIC.
Basically, snBox and epicBox are specified in
the config.  These are based on fractions of
the text rectangle -- likely to work across a
whole lot of lists.  If not, it would be possible
to tweak on an individual or global basis.

This is a far better method than what I had
previously.